+++ include/rodney/rodney_node.h
@@ -6,6 +6,7 @@
 #include <sensor_msgs/Joy.h>
 #include <sensor_msgs/BatteryState.h>
 #include <std_msgs/String.h>
+#include <std_msgs/Empty.h>
 #include <geometry_msgs/Twist.h>
 
 class RodneyNode
@@ -27,9 +28,11 @@ private:
     ros::Subscriber mission_sub_;       // Topic for mission complete indication
     ros::Subscriber demand_sub_;        // Topic for autonomous motor demands
     ros::Subscriber battery_status_sub_;// Topic to monitor battery status for the main battery
+    ros::Subscriber remote_heartbeat_sub_;  // Topic to monitor the remote heartbeat when teleoping
     ros::Time last_twist_send_time_;    // Time of previous message
     ros::Time last_battery_warn_;       // Time of the last spoken battery warning message
     ros::Time last_interaction_time_;   // Time that a human last interacted with the robot
+    ros::Time remote_heartbeat_time_;   // Time that the last remote heartbeat message was received
     
     geometry_msgs::Twist last_twist_;   // The last Twist message sent   
     
@@ -39,7 +42,7 @@ private:
     bool mission_running_;
     bool manual_locomotion_mode_;
     bool wav_play_enabled_;
-    bool pid_enabled_; 
+    bool pid_enabled_;
     
     unsigned int battery_low_count_; // Counter for low battery low messages   
  
@@ -48,7 +51,8 @@ private:
     int manual_mode_select_;   // Controller button index for manual mode (teleop)
     int camera_x_index_;       // Controller axes index for x direction head/camera
     int camera_y_index_;       // Controller axes index for y direction head/camera
-    int default_camera_pos_select_; // Controller button index formoving camera to default position
+    int default_camera_pos_select_; // Controller button index for moving camera to default position
+    int lidar_enable_select_;       // Controller button index for enable/disable LIDAR function
     
     int dead_zone_;                     // Controller dead zone value    
     
@@ -80,6 +84,7 @@ private:
     void completeCallBack(const std_msgs::String::ConstPtr& msg);
     void motorDemandCallBack(const geometry_msgs::Twist::ConstPtr& msg);
     void batteryCallback(const sensor_msgs::BatteryState::ConstPtr& msg);
+    void remHeartbeatCallback(const std_msgs::Empty::ConstPtr& msg);
     
     geometry_msgs::Twist rampedTwist(geometry_msgs::Twist prev, geometry_msgs::Twist target,
                                      ros::Time time_prev, ros::Time time_now);